Update a review if a user already reviewed a project, add test

+++ frux_app_server/graphqlschema/object.py
@@ -67,7 +67,7 @@ class Project(SQLAlchemyObjectType):
     amount_collected = graphene.Float()
     investor_count = graphene.Int()
     favorite_count = graphene.Int()
-    general_score = graphene.Int()
+    general_score = graphene.Float()
     review_count = graphene.Int()
 
     class Meta:
